Even more fixes to shifter
Fix to an odd page that could be mistaken as a start or end page, but is likely an end page. Added with the fact the first page is a spread, and the second is unknown and it gets funky.

source for these fixes: fix was applied for kanan-sama chap 40